Add movie poster and refactor code
A new MoviePoster component has been introduced to display a movie's poster on each MovieOverview. Additionally, movie types, importing paths, and Docker environment have been refactored for better organization and performance. Further, the moviesSampleData has been updated with changes to movie id and inclusion of a poster URL.
